Move config schema to schema folder
