+++ docs/hooks/generate_readable_schema.py
@@ -14,13 +14,13 @@ from pathlib import Path
 import jsonschema2md
 from mkdocs.structure.files import File
 
-from calliope import config
+from calliope.schemas import config_schema
 from calliope.util import schema
 
 TEMPDIR = tempfile.TemporaryDirectory()
 
 SCHEMAS = {
-    "config_schema": config.CalliopeConfig().model_no_ref_schema(),
+    "config_schema": config_schema.CalliopeConfig().model_no_ref_schema(),
     "model_schema": schema.MODEL_SCHEMA,
     "math_schema": schema.MATH_SCHEMA,
     "data_table_schema": schema.DATA_TABLE_SCHEMA,
