+++ src/calliope/backend/backend_model.py
@@ -26,12 +26,13 @@ from typing import (
 import numpy as np
 import xarray as xr
 
-from calliope import config, exceptions
+from calliope import exceptions
 from calliope.attrdict import AttrDict
 from calliope.backend import helper_functions, parsing
 from calliope.exceptions import warn as model_warn
 from calliope.io import load_config, to_yaml
 from calliope.preprocess.model_math import ORDERED_COMPONENTS_T, CalliopeMath
+from calliope.schemas import config_schema
 from calliope.util.schema import MODEL_SCHEMA, extract_from_schema
 
 if TYPE_CHECKING:
@@ -66,7 +67,7 @@ class BackendModelGenerator(ABC):
     _PARAM_TYPE = extract_from_schema(MODEL_SCHEMA, "x-type")
 
     def __init__(
-        self, inputs: xr.Dataset, math: CalliopeMath, build_config: config.Build
+        self, inputs: xr.Dataset, math: CalliopeMath, build_config: config_schema.Build
     ):
         """Abstract base class to build a representation of the optimisation problem.
 
@@ -606,7 +607,7 @@ class BackendModel(BackendModelGenerator, Generic[T]):
         self,
         inputs: xr.Dataset,
         math: CalliopeMath,
-        build_config: config.Build,
+        build_config: config_schema.Build,
         instance: T,
     ) -> None:
         """Abstract base class to build backend models that interface with solvers.
